+++ src/components/abstract-component.js
@@ -1,6 +1,9 @@
 import {createElement} from '../utils';
 export class AbstractComponent {
   constructor() {
+    if (new.target === AbstractComponent) {
+      throw new Error(`Can't instantiate AbstractComponent, only concrete one.`);
+    }
     this._element = null;
   }
